Level Dressing
- Added a new level to rigging
- added more verticality to Upper Deck
- Dressing.
Each room has its own set assets to quickly copy paste them to keep the hierarchy intact. Those are disabled for now.
- Rearranged Spawn Points to CC rooms.
This is temporary but it gives the room purpose for now
- Added additional Permanent Objects.
I'm not sure if the enemies will be able to reach the upper ones, making it impossible to loose the game. Gotta test it -_-

I feel good about this